requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==1.11.28
django-redis==4.12.1
get==2019.4.13
idna==2.9
post==2019.4.13
public==2019.4.13
pytz==2020.1
qcloudsms-py==0.1.4
query-string==2019.4.13
redis==3.5.3
request==2019.4.13
requests==2.23.0
sqlparse==0.3.1
tencentcloud-sdk-python==3.0.184
urllib3==1.25.9
